Create ExtraFunctions.js
ExtraFunctions.js for FirestoreApp
This file introduces additional utility functions to enhance the capabilities of the FirestoreApp library.

Features
getDocuments(collectionPath):
Retrieve all documents from a specified Firestore collection.
Parameters:
collectionPath (String): Path of the Firestore collection from which documents are to be fetched.
Returns:
An array of document data from the specified Firestore collection.
Usage
To use the functions provided by ExtraFunctions.js, make sure to import or reference it along with the main FirestoreApp library.

Contribution
Feel free to contribute to this file if you have suggestions for new utility functions or improvements to the existing ones. Please ensure that any added functions align with the theme and purpose of ExtraFunctions.js.